fix(analytics,images): GA4 back to afterInteractive, apostrophe-safe lazy guard
- GA4 moved from beforeInteractive (_document.tsx) back to afterInteractive
  in layout.tsx. afterInteractive fires right after hydration so we don't
  lose early pageviews (the lazyOnload problem #387 fixed), while keeping the
  485 KiB GTM script off the pre-hydration critical path so it doesn't
  compete with LCP. beforeInteractive bought a few ms at LCP's expense.
- Body-image lazy guard no longer counts raw quotes: an apostrophe inside a
  value (alt="Keploy's dashboard") wrongly tripped the bail, dropping the
  loading="lazy"/decoding="async" optimization. Now tokenizes attributes and
  bails only on leftover non-whitespace, so truncated tags are still skipped.
