+++ components/post-body.tsx
@@ -193,12 +193,17 @@ export default function PostBody({
       /<img\b([^>]*)>/gi,
       (match, attrs: string) => {
         // `[^>]*` stops at the first `>`, even one inside a quoted attribute
-        // (e.g. alt="a > b" from a raw Custom HTML block). If either quote type
-        // is unbalanced the match truncated mid-attribute, so leave it untouched
-        // rather than splicing attributes into the middle of a broken value.
-        const dq = (attrs.match(/"/g) || []).length;
-        const sq = (attrs.match(/'/g) || []).length;
-        if (dq % 2 !== 0 || sq % 2 !== 0) return match;
+        // (e.g. alt="a > b" from a raw Custom HTML block), which truncates the
+        // match mid-attribute. Detect that by consuming well-formed attributes
+        // (name, name="...", name='...', name=bare) and bailing only if any
+        // non-whitespace is left over — a leftover means a value was cut off.
+        // (Counting raw quotes would wrongly bail on a legit apostrophe inside
+        // a value, e.g. alt="Keploy's dashboard", dropping the optimization.)
+        const leftover = attrs.replace(
+          /\s+[^\s=]+(?:\s*=\s*(?:"[^"]*"|'[^']*'|[^\s"'>]+))?/g,
+          ""
+        );
+        if (leftover.trim() !== "") return match;
 
         // Strip a self-closing trailing slash so appends don't produce
         // "<img ... / loading=...>" (invalid markup).
